+++ addon/components/navbar-auth-dropdown/component.js
@@ -26,7 +26,7 @@ export default Ember.Component.extend({
     currentUser: Ember.inject.service(),
 
     tagName: 'li',
-    classNames: ['dropdown'],
+    classNames: ['dropdown', 'secondary-nav-dropdown'],
     classNameBindings: ['notAuthenticated:sign-in'],
     notAuthenticated: Ember.computed.not('session.isAuthenticated'),
     redirectUrl: null,