+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2018-05-29
+### Added
+- `scheduled-banner` component that pulls banners (created in the OSF Admin app) from the API
+
+### Changed
+- Format last edited date in search result like "MMM DDD, YYYY UTC" instead of "YYYY-MM-DD (UTC)"
+- `navbar-auth-dropdown` to match current changes to the navbar
+
+### Fixed
+- Missing banner translation error
+
 ## [0.16.2] - 2018-05-01
 ### Added
 - temporary inline style to sign-up button
@@ -25,6 +36,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OSF API version from 2.4 to 2.6
 - ember-cli-moment-shim version to `^3.5.3` due to security issues found in `moment` versions before `2.19.3`
 - classes and element tags to match styles needed for long name truncation on navbar
+
+### Added
+- Add moderator model, serializer, and adapter
 
 ## [0.15.0] - 2018-03-06
 ### Added